chore: rename Common to Nostr
